// chat_rooms_client/src/api/models/message.ts
/* tslint:disable */
/* eslint-disable */
/**
 * ChatRoomsWeb
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * OpenAPI spec version: 1.0
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */

 /**
 * 
 *
 * @export
 * @interface Message
 */
export interface Message {

    /**
     * @type {string}
     * @memberof Message
     */
    id?: string | null;

    /**
     * @type {string}
     * @memberof Message
     */
    chatRoomId?: string | null;

    /**
     * @type {string}
     * @memberof Message
     */
    username?: string | null;

    /**
     * @type {Date}
     * @memberof Message
     */
    dateTimeSent?: Date;

    /**
     * @type {string}
     * @memberof Message
     */
    text?: string | null;
}